docs: document the deployment builder interface
Move the interface out of internal and under `./types/dsl.ts`.

Add typedocs for each function on the interface to better support
intellisense.

All parts of this api are intended for public use - hence the @alpha tag.
