+++ packages/core/src/dsl/buildModule.ts
@@ -1,10 +1,10 @@
-import type { IDeploymentBuilder } from "../internal/types/deploymentGraph";
 import type { Module, ModuleDict } from "../types/module";
 
 import {
   assertFunctionParam,
   assertStringParam,
 } from "../internal/utils/paramAssertions";
+import { IDeploymentBuilder } from "../types/dsl";
 
 /**
  * A factory that builds a deployment module given configuration function that
