+++ packages/core/test/deploymentBuilder/artifacts.ts
@@ -1,14 +1,12 @@
 /* eslint-disable import/no-unused-modules */
-import type {
-  IDeploymentBuilder,
-  IDeploymentGraph,
-} from "../../src/internal/types/deploymentGraph";
+import type { IDeploymentGraph } from "../../src/internal/types/deploymentGraph";
 
 import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
 import { isArtifactContract } from "../../src/internal/utils/guards";
+import { IDeploymentBuilder } from "../../src/types/dsl";
 
 import { getDeploymentVertexByLabel } from "./helpers";
