+++ packages/core/test/deploymentBuilder/metadata.ts
@@ -3,7 +3,7 @@ import { assert } from "chai";
 
 import { buildModule } from "../../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
-import { IDeploymentBuilder } from "../../src/internal/types/deploymentGraph";
+import { IDeploymentBuilder } from "../../src/types/dsl";
 
 describe("deployment builder - metadata", () => {
   it("should inject the chainId via the builder", () => {
