+++ packages/core/test/deploymentBuilder/sendETH.ts
@@ -4,11 +4,9 @@ import { ethers } from "ethers";
 
 import { buildModule } from "../../src/dsl/buildModule";
 import { generateDeploymentGraphFrom } from "../../src/internal/process/generateDeploymentGraphFrom";
-import {
-  IDeploymentBuilder,
-  IDeploymentGraph,
-} from "../../src/internal/types/deploymentGraph";
+import { IDeploymentGraph } from "../../src/internal/types/deploymentGraph";
 import { isHardhatContract } from "../../src/internal/utils/guards";
+import { IDeploymentBuilder } from "../../src/types/dsl";
 
 import {
   getDependenciesForVertex,
